# docs/agent/openapi.yaml
openapi: 3.1.0
info:
  title: BoardAgent API
  description: 'BoardAgent local task manager API. All endpoints except /healthz require
    the X-API-Key header (roles: read < write < admin). See docs/agent/rest_api.md
    for the full error table.'
  version: 0.1.0
paths:
...
  /settings:
    get:
      summary: Get Settings
      operationId: get_settings_settings_get
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSettings'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Put Settings
      operationId: put_settings_settings_put
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerSettings'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSettings'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /keys:
    get:
      summary: List Keys
      operationId: list_keys_keys_get
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
                title: Response List Keys Keys Get
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Key
      operationId: create_key_keys_post
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /keys/{key}:
    delete:
      summary: Delete Key
      operationId: delete_key_keys__key__delete
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          title: Key
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Successful Response
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '400':
          description: Bad transition or invalid request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks:
    post:
      summary: Create Task
      operationId: create_task_tasks_post
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Tasks
      operationId: list_tasks_tasks_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Status'
          - type: 'null'
          title: Status
      - name: project
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Project
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 500
            minimum: 1
          - type: 'null'
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}:
    get:
      summary: Get Task
      operationId: get_task_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Task
      operationId: update_task_tasks__task_id__patch
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '400':
          description: Bad transition or invalid request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Task
      operationId: delete_task_tasks__task_id__delete
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Successful Response
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/claim:
    post:
      summary: Claim Task
      operationId: claim_task_tasks__task_id__claim_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskClaim'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '409':
          description: Already claimed / already done
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/complete:
    post:
      summary: Complete Task
      operationId: complete_task_tasks__task_id__complete_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskComplete'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '404':
          description: Task or key not found
        '409':
          description: Already claimed / already done
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/schema/priority:
    get:
      summary: Priority Values
      operationId: priority_values_tasks_schema_priority_get
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                title: Response Priority Values Tasks Schema Priority Get
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/schema/status:
    get:
      summary: Status Values
      operationId: status_values_tasks_schema_status_get
      parameters:
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                title: Response Status Values Tasks Schema Status Get
        '401':
          description: Missing or invalid API key
        '403':
          description: Insufficient role or not the owner
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ApiKey:
      properties:
        key:
          type: string
          title: Key
        name:
          type: string
          title: Name
        role:
          $ref: '#/components/schemas/Role'
      type: object
      required:
      - key
      - name
      - role
      title: ApiKey
    ApiKeyCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        role:
          $ref: '#/components/schemas/Role'
          default: read
      type: object
      required:
      - name
      title: ApiKeyCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ServerSettings:
      properties:
        host:
          type: string
          maxLength: 255
          title: Host
          default: 127.0.0.1
        port:
          type: integer
          maximum: 65535.0
          minimum: 1.0
          title: Port
          default: 7373
        api_enabled:
          type: boolean
          title: Api Enabled
          default: true
        mcp_enabled:
          type: boolean
          title: Mcp Enabled
          default: true
      additionalProperties: false
      type: object
      title: ServerSettings
    Status:
      type: string
      enum:
      - todo
      - in_progress
      - blocked
      - done
      title: Status
    Task:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        due:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due
        priority:
          $ref: '#/components/schemas/Priority'
          default: white
        project:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Project
        status:
          $ref: '#/components/schemas/Status'
          default: todo
        tags:
          items:
            type: string
          type: array
          title: Tags
        estimate:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Estimate
        links:
          items:
            type: string
          type: array
          title: Links
        acceptance_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Acceptance Criteria
        dependencies:
          items:
            type: string
          type: array
          title: Dependencies
        notes:
          items:
            type: string
          type: array
          title: Notes
        custom_fields:
          additionalProperties:
            type: string
          type: object
          title: Custom Fields
        id:
          type: integer
          title: Id
        owner_agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Agent Id
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - title
      - id
      - created_at
      - updated_at
      title: Task
    TaskClaim:
      properties:
        agent_id:
          type: string
          maxLength: 100
          minLength: 1
          title: Agent Id
      type: object
      required:
      - agent_id
      title: TaskClaim
    TaskComplete:
      properties:
        agent_id:
          type: string
          maxLength: 100
          minLength: 1
          title: Agent Id
      type: object
      required:
      - agent_id
      title: TaskComplete
    TaskCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        due:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due
        priority:
          $ref: '#/components/schemas/Priority'
          default: white
        project:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Project
        status:
          $ref: '#/components/schemas/Status'
          default: todo
        tags:
          items:
            type: string
          type: array
          title: Tags
        estimate:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Estimate
        links:
          items:
            type: string
          type: array
          title: Links
        acceptance_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Acceptance Criteria
        dependencies:
          items:
            type: string
          type: array
          title: Dependencies
        notes:
          items:
            type: string
          type: array
          title: Notes
        custom_fields:
          additionalProperties:
            type: string
          type: object
          title: Custom Fields
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
        agent_id:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Agent Id
      additionalProperties: false
      type: object
      required:
      - title
      title: TaskCreate
    TaskList:
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/Task'
          type: array
          title: Tasks
        count:
          type: integer
          title: Count
      type: object
      required:
      - tasks
      - count
      title: TaskList
...